@@ -7,11 +7,11 @@ license = "MIT"
 readme = ["README.md", "LICENSE"]
 
 [tool.poetry.dependencies]
-python = ">=3.9, <3.13"
-pydantic = "^1.10.7"
+python = "^3.10"
+pydantic = "^2.7.1"
 simpy = "^4.0.1"
-fastapi = "0.99.1"
-uvicorn = {extras = ["standard"], version = "^0.21.1"}
+fastapi = "^0.111.0"
+uvicorn = {extras = ["standard"], version = "^0.29.0"}
 deap = "^1.3.3"
 simanneal = "^0.5.0"
 gurobipy = "^10.0.1"
