@@ -6,12 +6,12 @@ from typing import Optional
 from uuid import uuid1
 
 
-from pydantic import Field, conlist
+from pydantic import ConfigDict, Field, conlist
 from pydantic.dataclasses import dataclass
 
-from prodsys.models import node_data
-
 from prodsys.express import core
+
+from prodsys.models import node_data
 
 @dataclass
 class Node(core.ExpressObject):
@@ -21,20 +21,18 @@ class Node(core.ExpressObject):
         Attributes:
                 location (List[float]): Location of the node. It has to be a list of length 2.
         """
-        location: conlist(float, min_items=2, max_items=2) # type: ignore
+        location: conlist(float, min_length=2, max_length=2) # type: ignore
         ID: Optional[str] = Field(default_factory=lambda: str(uuid1()))
 
-        class Config:
-                schema_extra = {
-                "example": {
-                        "summary": "Node",
-                        "value": {
-                        "ID": "N1",
-                        "description": "Node 1",
-                        "location": [0.0, 0.0],
-                        },
-                }
-            }
+        model_config=ConfigDict(json_schema_extra= {
+                "examples": [
+                        {
+                                "ID": "N1",
+                                "description": "Node 1",
+                                "location": [0.0, 0.0],
+                        }
+                ]
+        })
                 
         def to_model(self) -> node_data.NodeData:
              """
