@@ -1,9 +1,7 @@
-from __future__ import annotations
-
 from enum import Enum
 from typing import Literal, List, Optional, Union, Tuple
 
-from pydantic import BaseModel, validator
+from pydantic import BaseModel, ConfigDict, field_validator
 
 
 class KPIEnum(str, Enum):
@@ -63,7 +61,7 @@ class KPI(BaseModel):
     resource: Optional[str] = None
     product_type: Optional[str] = None
 
-    @validator("context")
+    @field_validator("context", mode="before")
     def sort_context(cls, v):
         return tuple(sorted(v))
 
@@ -83,243 +81,228 @@ class Output(KPI):
     name: Literal[KPIEnum.OUTPUT] = KPIEnum.OUTPUT
     target: Literal["max"] = "max"
 
-    class Config:
-        schema_extra = {
-            "example": {
-                "summary": "Output KPI",
-                "value": {
+    model_config = ConfigDict(
+        json_schema_extra={
+            "examples": [
+                {
                     "name": "output",
                     "target": "max",
                     "weight": 1,
                     "value": 34,
                     "context": ["system", "product_type"],
                     "product_type": "ProductType_1",
-                },
-            }
+                }
+            ]
         }
+    )
 
 
 class Throughput(KPI):
     name: Literal[KPIEnum.THROUGHPUT]
     target: Literal["max"] = "max"
 
-    class Config:
-        schema_extra = {
-            "example": {
-                "summary": "Throughput KPI",
-                "value": {
+    model_config = ConfigDict(
+        json_schema_extra={
+            "examples": [
+                {
                     "name": "throughput",
                     "target": "max",
                     "weight": 1,
                     "value": 4.32,
                     "context": ["system", "product_type"],
                     "product_type": "ProductType_1",
-                },
-            }
+                }
+            ]
         }
-
+    )
 
 class Cost(KPI):
     name: Literal[KPIEnum.COST]
     target: Literal["min"] = "min"
 
-    class Config:
-        schema_extra = {
-            "example": {
-                "summary": "Cost KPI",
-                "value": {
+    model_config = ConfigDict(
+        json_schema_extra={
+            "examples": [
+                {
                     "name": "cost",
                     "target": "min",
                     "weight": 0.5,
                     "value": 36000,
                     "context": ["system"],
-                },
-            }
+                }
+            ]
         }
+    )
 
 
 class WIP(KPI):
     name: Literal[KPIEnum.WIP]
     target: Literal["min"] = "min"
 
-    class Config:
-        schema_extra = {
-            "example": {
-                "summary": "WIP KPI",
-                "value": {
+    model_config = ConfigDict(
+        json_schema_extra={
+            "examples": [
+                {
                     "name": "WIP",
                     "target": "min",
                     "weight": 1,
                     "value": 121,
                     "context": ["system", "product_type"],
                     "product_type": "ProductType_1",
-                },
-            }
+                }
+            ]
         }
+    )
 
 
 class DynamicWIP(DynamicKPI, WIP):
     name: Literal[KPIEnum.DYNAMIC_WIP]
     target: Literal["min"] = "min"
 
-    class Config:
-        schema_extra = {
-            "example": {
-                "summary": "Dynamic WIP KPI",
-                "value": {
-                    "name": "dynamic_WIP",
-                    "target": "min",
-                    "weight": 1,
-                    "value": 121,
-                    "context": ["system", "product"],
-                    "product_type": "ProductType_1",
-                    "start_time": 21.2,
-                    "end_time": 23.4,
-                },
+    model_config = ConfigDict(json_schema_extra={
+        "examples": [
+            {
+                "name": "dynamic_WIP",
+                "target": "min",
+                "weight": 1,
+                "value": 121,
+                "context": ["system", "product"],
+                "product_type": "ProductType_1",
+                "start_time": 21.2,
+                "end_time": 23.4,
             }
-        }
+        ]
+    })
 
 
 class ThroughputTime(KPI):
     name: Literal[KPIEnum.TRHOUGHPUT_TIME]
     target: Literal["min"] = "min"
 
-    class Config:
-        schema_extra = {
-            "example": {
-                "summary": "Throughput time KPI",
-                "value": {
-                    "name": "throughput_time",
-                    "target": "min",
-                    "weight": 1,
-                    "value": 221.1,
-                    "context": ["system", "product_type"],
-                    "product_type": "ProductType_1",
-                },
+    model_config = ConfigDict(json_schema_extra={
+        "examples": [
+            {
+                "name": "throughput_time",
+                "target": "min",
+                "weight": 1,
+                "value": 221.1,
+                "context": ["system", "product_type"],
+                "product_type": "ProductType_1",
             }
-        }
+        ]
+    })
 
 
 class DynamicThroughputTime(DynamicKPI, ThroughputTime):
     name: Literal[KPIEnum.DYNAMIC_THROUGHPUT_TIME]
     target: Literal["min"] = "min"
 
-    class Config:
-        schema_extra = {
-            "example": {
-                "summary": "Dynamic throughput time KPI",
-                "value": {
-                    "name": "throughput_time",
-                    "target": "min",
-                    "weight": 1,
-                    "value": 201.3,
-                    "context": ["system", "product"],
-                    "product_type": "ProductType_1",
-                    "product": "Product_1_23",
-                },
+    model_config = ConfigDict(json_schema_extra={
+        "examples": [
+            {
+                "name": "dynamic_throughput_time",
+                "target": "min",
+                "weight": 1,
+                "value": 221.1,
+                "context": ["system", "product"],
+                "product_type": "ProductType_1",
+                "product": "Product_1_23",
+                "start_time": 21.2,
+                "end_time": 23.4,
             }
-        }
+        ]
+    })
 
 
 class ProcessingTime(KPI):
     name: Literal[KPIEnum.PROCESSING_TIME]
     target: Literal["min"] = "min"
 
-    class Config:
-        schema_extra = {
-            "example": {
-                "summary": "Processing time KPI",
-                "value": {
-                    "name": "processing_time",
-                    "target": "min",
-                    "weight": 1,
-                    "value": 1.2,
-                    "context": ["resource", "process"],
-                    "resource": "Resource_1",
-                    "process": "P1",
-                },
+    model_config = ConfigDict(json_schema_extra={
+        "examples": [
+            {
+                "name": "processing_time",
+                "target": "min",
+                "weight": 1,
+                "value": 1.2,
+                "context": ["resource", "process"],
+                "resource": "Resource_1",
+                "process": "P1",
             }
-        }
+        ]
+    })
 
 
 class ProductiveTime(KPI):
     name: Literal[KPIEnum.PRODUCTIVE_TIME]
     target: Literal["max"] = "max"
 
-    class Config:
-        schema_extra = {
-            "example": {
-                "summary": "Productive time KPI",
-                "value": {
-                    "name": "productive_time",
-                    "target": "max",
-                    "weight": 1,
-                    "value": 0.65,
-                    "context": ["resource"],
-                    "resource": "Resource_1",
-                },
+    model_config = ConfigDict(json_schema_extra={
+        "examples": [
+            {
+                "name": "productive_time",
+                "target": "max",
+                "weight": 1,
+                "value": 0.65,
+                "context": ["resource"],
+                "resource": "Resource_1",
             }
-        }
+        ]
+    })
 
 
 class StandbyTime(KPI):
     name: Literal[KPIEnum.STANDBY_TIME]
     target: Literal["min"] = "min"
 
-    class Config:
-        schema_extra = {
-            "example": {
-                "summary": "Standby time KPI",
-                "value": {
-                    "name": "standby_time",
-                    "target": "min",
-                    "weight": 1,
-                    "value": 0.12,
-                    "context": ["resource"],
-                    "resource": "Resource_1",
-                },
+    model_config = ConfigDict(json_schema_extra={
+        "examples": [
+            {
+                "name": "standby_time",
+                "target": "min",
+                "weight": 1,
+                "value": 0.12,
+                "context": ["resource"],
+                "resource": "Resource_1",
             }
-        }
+        ]
+    })
 
 
 class SetupTime(KPI):
     name: Literal[KPIEnum.SETUP_TIME]
     target: Literal["min"] = "min"
 
-    class Config:
-        schema_extra = {
-            "example": {
-                "summary": "Setup time KPI",
-                "value": {
-                    "name": "setup_time",
-                    "target": "min",
-                    "weight": 1,
-                    "value": 0.08,
-                    "context": ["resource"],
-                    "resource": "Resource_1",
-                },
+    model_config = ConfigDict(json_schema_extra={
+        "examples": [
+            {
+                "name": "setup_time",
+                "target": "min",
+                "weight": 1,
+                "value": 0.08,
+                "context": ["resource"],
+                "resource": "Resource_1",
             }
-        }
+        ]
+    })
 
 
 class UnscheduledDowntime(KPI):
     name: Literal[KPIEnum.UNSCHEDULED_DOWNTIME]
     target: Literal["min"] = "min"
 
-    class Config:
-        schema_extra = {
-            "example": {
-                "summary": "Unscheduled downtime KPI",
-                "value": {
-                    "name": "unscheduled_downtime",
-                    "target": "min",
-                    "weight": 1,
-                    "value": 0.1,
-                    "context": ["resource"],
-                    "resource": "Resource_1",
-                },
+    model_config = ConfigDict(json_schema_extra={
+        "examples": [
+            {
+                "name": "unscheduled_downtime",
+                "target": "min",
+                "weight": 1,
+                "value": 0.1,
+                "context": ["resource"],
+                "resource": "Resource_1",
             }
-        }
+        ]
+    })
 
 
 KPI_UNION = Union[
