@@ -6,13 +6,10 @@ This module contains the data structures for the scenario data that is used in o
 - 'Objectives': The objectives of the scenario.
 - `ScenarioData`: The scenario data that contains the constraints, options, information and objectives of the scenario. 
 """
-
-from __future__ import annotations
-
 from typing import Literal, List, Optional, Dict
 from enum import Enum
 
-from pydantic import BaseModel, validator, conlist
+from pydantic import BaseModel, ConfigDict, field_validator, conlist
 
 from prodsys.models.performance_indicators import KPIEnum
 
@@ -57,20 +54,17 @@ class ScenarioConstrainsData(BaseModel):
     max_num_transport_resources: int
     target_product_count: Optional[Dict[str, int]]
 
-    class Config:
-        schema_extra = {
-            "example": {
-                "summary": "Scenario constraints",
-                "value": {
-                    "max_reconfiguration_cost": 120000,
-                    "max_num_machines": 10,
-                    "max_num_processes_per_machine": 2,
-                    "max_num_transport_resources": 2,
-                    "target_product_count": {"Product_1": 120, "Product_2": 200},
-                },
+    model_config=ConfigDict(json_schema_extra={
+        "examples": [
+            {
+                "max_reconfiguration_cost": 120000,
+                "max_num_machines": 10,
+                "max_num_processes_per_machine": 2,
+                "max_num_transport_resources": 2,
+                "target_product_count": {"Product_1": 120, "Product_2": 200},
             }
-        }
-
+        ]
+    })
 
 class ScenarioOptionsData(BaseModel):
     """
@@ -84,7 +78,7 @@ class ScenarioOptionsData(BaseModel):
         machine_controllers (List[Literal["FIFO", "LIFO", "SPT"]]): List of possible controllers for machines.
         transport_controllers (List[Literal["FIFO", "SPT_transport", "Nearest_origin_and_longest_target_queues_transport", "Nearest_origin_and_shortest_target_input_queues_transport"]]): List of possible controllers for transport resources.
         routing_heuristics (List[Literal["shortest_queue", "random", "FIFO"]]): List of possible routing heuristics for sources.
-        positions (List[conlist(float, min_items=2, max_items=2)]): List of possible positions for machines in the layout.
+        positions (List[conlist(float, min_length=2, max_length=2)]): List of possible positions for machines in the layout.
 
     Raises:
         ValueError: If the positions are not a list of tuples of length 2.
@@ -94,27 +88,32 @@ class ScenarioOptionsData(BaseModel):
     machine_controllers: List[Literal["FIFO", "LIFO", "SPT"]]
     transport_controllers: List[Literal["FIFO", "SPT_transport"]]
     routing_heuristics: List[Literal["shortest_queue", "random", "FIFO"]]
-    positions: List[conlist(float, min_items=2, max_items=2)] # type: ignore
+    positions: List[conlist(float, min_length=2, max_length=2)] # type: ignore
 
-    @validator("positions")
+    @field_validator("positions")
     def check_positions(cls, v):
         for e in v:
             if len(e) != 2:
                 raise ValueError("positions must be a list of tuples of length 2")
         return v
 
-    class Config:
-        schema_extra = {
-            "example": {
-                "summary": "Scenario options",
-                "value": {
-                    "machine_controllers": ["FIFO", "LIFO", "SPT"],
-                    "transport_controllers": ["FIFO", "SPT_transport"],
-                    "routing_heuristics": ["shortest_queue", "random", "FIFO"],
-                    "positions": [[10.0, 10.0], [20.0, 20.0]],
-                },
+    model_config=ConfigDict(json_schema_extra={
+        "examples": [
+            {
+                "transformations": [
+                    "production_capacity",
+                    "transport_capacity",
+                    "layout",
+                    "sequencing_logic",
+                    "routing_logic",
+                ],
+                "machine_controllers": ["FIFO", "LIFO", "SPT"],
+                "transport_controllers": ["FIFO", "SPT_transport"],
+                "routing_heuristics": ["shortest_queue", "random", "FIFO"],
+                "positions": [[10.0, 10.0], [20.0, 20.0]],
             }
-        }
+        ]
+    })
 
 
 class ScenarioInfoData(BaseModel):
@@ -138,43 +137,40 @@ class ScenarioInfoData(BaseModel):
     time_range: Optional[int]
     maximum_breakdown_time: Optional[int]
 
-    class Config:
-        schema_extra = {
-            "example": {
-                "summary": "Scenario information",
-                "value": {
-                    "machine_cost": 30000,
-                    "transport_resource_cost": 20000,
-                    "process_module_cost": 2300,
-                    "breakdown_cost": 1000,
-                    "time_range": 2600,
-                    "maximum_breakdown_time": 10,
-                },
+    model_config=ConfigDict(json_schema_extra={
+        "examples": [
+            {
+                "machine_cost": 30000,
+                "transport_resource_cost": 20000,
+                "process_module_cost": 2300,
+                "breakdown_cost": 1000,
+                "time_range": 2600,
+                "maximum_breakdown_time": 10,
             }
-        }
+        ]
+    })
 
 
 class Objective(BaseModel):
     name: KPIEnum
     weight: float = 1.0
 
-    class Config:
-        schema_extra = {
-            "examples": [
-                {
-                    "name": KPIEnum.COST,
-                    "weight": 0.6,
-                },
-                {
-                    "name": KPIEnum.THROUGHPUT,
-                    "weight": 0.1,
-                },
-                {
-                    "name": KPIEnum.WIP,
-                    "weight": 0.5,
-                },
-            ]
-        }
+    model_config=ConfigDict(json_schema_extra={
+        "examples": [
+            {
+                "name": KPIEnum.COST,
+                "weight": 0.6,
+            },
+            {
+                "name": KPIEnum.THROUGHPUT,
+                "weight": 0.1,
+            },
+            {
+                "name": KPIEnum.WIP,
+                "weight": 0.5,
+            },
+        ]
+    })
 
 
 class ScenarioData(BaseModel):
@@ -196,10 +192,9 @@ class ScenarioData(BaseModel):
     info: ScenarioInfoData
     objectives: List[Objective]
 
-    class Config:
-        use_enum_values = True
-        schema_extra = {
-            "example": {
+    model_config=ConfigDict(use_enum_values=True, json_schema_extra={
+        "examples": [
+            {
                 "summary": "Scenario",
                 "value": {
                     "constraints": {
@@ -246,4 +241,5 @@ class ScenarioData(BaseModel):
                     ],
                 },
             }
-        }
+        ]
+    })
