@@ -20,6 +20,7 @@ The following modules are available:
 - `prodsys.express.time_model`: Contains classes to specify time models.
 
 """
+from prodsys.express.core import ExpressObject
 from prodsys.express.time_model import (
     SampleTimeModel,
     ScheduledTimeModel,
@@ -28,6 +29,8 @@ from prodsys.express.time_model import (
     SequentialTimeModel,
     ManhattanDistanceTimeModel,
 )
+from prodsys.express.state import SetupState, BreakDownState, ProcessBreakdownState
+from prodsys.express.node import Node
 from prodsys.express.process import (
     ProductionProcess,
     CapabilityProcess,
@@ -35,9 +38,7 @@ from prodsys.express.process import (
     TransportProcess,
     LinkTransportProcess,
 )
-from prodsys.express.state import SetupState, BreakDownState, ProcessBreakdownState
 from prodsys.express.resources import ProductionResource, TransportResource
-from prodsys.express.node import Node
 from prodsys.express.product import Product
 from prodsys.express.source import Source
 from prodsys.express.sink import Sink
