@@ -2,29 +2,27 @@ from __future__ import annotations
 
 from typing import List, TYPE_CHECKING
 
-from pydantic import BaseModel, parse_obj_as
+from pydantic import BaseModel, ConfigDict, TypeAdapter
 
-from prodsys.simulation import sim
 
 from prodsys.models.node_data import NodeData
 from prodsys.simulation import node
 
 if TYPE_CHECKING:
     from prodsys.adapters import adapter
+    from prodsys.simulation import sim
 
-class NodeFactory(BaseModel):
+class NodeFactory:
     """
     Factory class that creates and stores `prodsys.simulation` resource objects from `prodsys.models` node objects.
 
     Args:
         env (sim.Environment): prodsys simulation environment.
     """
-    env: sim.Environment
-    nodes: List[node.Node] = []
-    
 
-    class Config:
-        arbitrary_types_allowed = True
+    def __init__(self, env: sim.Environment):
+        self.env = env
+        self.nodes = []
 
     def create_nodes(self, adapter: adapter.ProductionSystemAdapter):
         """
@@ -45,7 +43,7 @@ class NodeFactory(BaseModel):
         """
         values = {}
         values.update({"data": node_data})
-        self.nodes.append(parse_obj_as(node.Node, values))
+        self.nodes.append(TypeAdapter(node.Node).validate_python(values))
 
     def get_node(self, ID: str) -> node.Node:
         """
