@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import List, Optional, TYPE_CHECKING
 
-from pydantic import BaseModel, parse_obj_as
+from pydantic import BaseModel, TypeAdapter
 
 from prodsys.factories import time_model_factory
 from prodsys.models import processes_data
@@ -38,15 +38,18 @@ class ProcessFactory(BaseModel):
         if not (isinstance(process_data, processes_data.CompoundProcessData) or isinstance(process_data, processes_data.RequiredCapabilityProcessData)):
             time_model = self.time_model_factory.get_time_model(process_data.time_model_id)
             values.update({"time_model": time_model})
+        else:
+            values.update({"time_model": None})
         values.update({"process_data": process_data})
         if isinstance(process_data, processes_data.CompoundProcessData):
             contained_processes_data = [other_process_data for other_process_data in adapter.process_data if other_process_data.ID in process_data.process_ids]
             values.update({"contained_processes_data": contained_processes_data})
         if isinstance(process_data, processes_data.LinkTransportProcessData):
             values.update({"links": [[]]})
-            self.processes.append(parse_obj_as(process.LinkTransportProcess, values))
+            self.processes.append(TypeAdapter(process.LinkTransportProcess).validate_python(values))
         else:
-            self.processes.append(parse_obj_as(process.PROCESS_UNION, values))
+            print("validate process", type(process_data))
+            self.processes.append(TypeAdapter(process.PROCESS_UNION).validate_python(values))
 
     def get_processes_in_order(self, IDs: List[str]) -> List[process.PROCESS_UNION]:
         """
@@ -86,4 +89,3 @@ class ProcessFactory(BaseModel):
 
 
 from prodsys.simulation import process
-ProcessFactory.update_forward_refs()
