@@ -1,7 +1,5 @@
-from __future__ import annotations
-
 from enum import Enum
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict
 from typing import Literal, Union, Optional, List, Tuple, TYPE_CHECKING
 
 from prodsys.models.performance_indicators import (
@@ -9,9 +7,6 @@ from prodsys.models.performance_indicators import (
     KPILevelEnum,
     KPI_UNION,
 )
-
-if TYPE_CHECKING:
-    from prodsys.simulation import state
 
 
 class Event(BaseModel):
@@ -38,31 +33,30 @@ class Event(BaseModel):
     expected_end_time: Optional[float] = None
     target_location: Optional[str] = None
 
-    class Config:
-        schema_extra = {
-            "examples": [
-                {
-                    "time": 12.0,
-                    "resource": "R1",
-                    "state": "P1",
-                    "state_type": "Production",
-                    "activity": "start state",
-                    "product": "Product_1_12",
-                    "expected_end_time": 24.3,
-                    "target_location": None,
-                }, 
-                {
-                    "time": 24.3,
-                    "resource": "R1",
-                    "state": "P1",
-                    "state_type": "Production",
-                    "activity": "end state",
-                    "product": "Product_1_12",
-                    "expected_end_time": None,
-                    "target_location": "L1",
-                },
-            ]
-        }
+    model_config = ConfigDict(json_schema_extra={
+        "examples": [
+            {
+                "time": 12.0,
+                "resource": "R1",
+                "state": "P1",
+                "state_type": "Production",
+                "activity": "start state",
+                "product": "Product_1_12",
+                "expected_end_time": 24.3,
+                "target_location": None,
+            }, 
+            {
+                "time": 24.3,
+                "resource": "R1",
+                "state": "P1",
+                "state_type": "Production",
+                "activity": "end state",
+                "product": "Product_1_12",
+                "expected_end_time": None,
+                "target_location": "L1",
+            },
+        ]
+    })
 
 
 class Performance(BaseModel):
@@ -74,31 +68,31 @@ class Performance(BaseModel):
         kpis (List[KPI_UNION]): List of KPIs of the simulation run.
     """
 
-    class Config:
-        schema_extra = {
-            "examples": [{               
-                    "event_log": Event.Config.schema_extra["examples"],
-                    "kpis": [
-                        {
-                            "name": "throughput",
-                            "target": "max",
-                            "weight": 1,
-                            "value": 4.32,
-                            "context": ["system", "product_type"],
-                            "product_type": "ProductType_1",
-                        },
-                        {
-                            "name": "WIP",
-                            "target": "min",
-                            "weight": 1,
-                            "value": 121,
-                            "context": ["system", "product_type"],
-                            "product_type": "ProductType_1",
-                        },
-                    ],
-                }
-            ]
-        }
+    model_config = ConfigDict(json_schema_extra={
+        "examples": [
+            {
+                "event_log": Event.model_config["json_schema_extra"]["examples"],
+                "kpis": [
+                    {
+                        "name": "throughput",
+                        "target": "max",
+                        "weight": 1,
+                        "value": 4.32,
+                        "context": ["system", "product_type"],
+                        "product_type": "ProductType_1",
+                    },
+                    {
+                        "name": "WIP",
+                        "target": "min",
+                        "weight": 1,
+                        "value": 121,
+                        "context": ["system", "product_type"],
+                        "product_type": "ProductType_1",
+                    },
+                ],
+            }
+        ]
+    })
 
     event_log: List[Event]
     kpis: List[KPI_UNION]
