@@ -1,7 +1,7 @@
-from __future__ import annotations
-
 from typing import Union
 from hashlib import md5
+
+from pydantic import ConfigDict
 
 from prodsys.models.core_asset import CoreAsset
 
@@ -48,19 +48,18 @@ class QueueData(CoreAsset):
         """
         return md5((str(self.capacity)).encode("utf-8")).hexdigest()
 
-
-    class Config:
-        schema_extra = {
-            "examples": [
-                {
-                    "ID": "Q1",
-                    "description": "Finte Queue",
-                    "capacity": 10,
-                },
-                {
-                    "ID": "Q1",
-                    "description": "Infinite Queue",
-                    "capacity": 0.0,
-                },
-            ]
-        }
+    model_config=ConfigDict(json_schema_extra={
+        "examples": [
+            {
+                "ID": "Q1",
+                "description": "Finte Queue",
+                "capacity": 10,
+            },
+            {
+                "ID": "Q1",
+                "description": "Infinite Queue",
+                "capacity": 0.0,
+            },
+        ]
+    
+    })
