@@ -12,8 +12,8 @@ import numpy as np
 
 from simpy import events
 
-from prodsys.simulation import resources
 from prodsys.simulation import request
+from prodsys.simulation import resources
 
 
 if TYPE_CHECKING:
